satvik-ai/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
lightrag-hku==1.3.5
openai==1.54.0
neo4j==5.27.0
psycopg2-binary==2.9.9
pgvector==0.3.6
asyncpg==0.30.0
python-dotenv==1.0.1
